inserted Amount in ingredients table

+++ app/Http/Controllers/ProjectDetailController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\ProjectDetail;
 use App\Models\ProjectDetailStyle;
+use App\Account;
 
 class ProjectDetailController extends Controller
 {
@@ -29,24 +30,29 @@ class ProjectDetailController extends Controller
 
     public function update(Request $request, $id)
     {
-        $projectdetailstyles = ProjectDetailStyle::where('projectdetailid', $id)->get();
-        foreach($projectdetailstyles as $projectdetailstyle) $projectdetailstyle->delete();
+			$projectdetail = $request->all();
+      if(Account::checkOwner($projectdetail['projectid'])==false && Account::checkTeamManager($id)==false) return;
 
-        $projectdetail = $request->all();
-        $projectdetailstyles = $projectdetail['projectdetailstyles'];
-        unset($projectdetail['projectdetailstyles']);
-        $res = ProjectDetail::find($id)->update($projectdetail);
-        foreach($projectdetailstyles as $projectdetailstyle){
-            $projectdetailstyle['projectdetailid'] = $projectdetail['id'];
-            $projectdetailstyle = ProjectDetailStyle::create($projectdetailstyle);
-        }
-        $projectdetail['projectdetailstyles'] = $projectdetailstyles;
+			$projectdetailstyles = ProjectDetailStyle::where('projectdetailid', $id)->get();
+			foreach($projectdetailstyles as $projectdetailstyle) $projectdetailstyle->delete();
 
-        return $projectdetail;
+
+			$projectdetailstyles = $projectdetail['projectdetailstyles'];
+			unset($projectdetail['projectdetailstyles']);
+			$res = ProjectDetail::find($id)->update($projectdetail);
+			foreach($projectdetailstyles as $projectdetailstyle){
+					$projectdetailstyle['projectdetailid'] = $projectdetail['id'];
+					$projectdetailstyle = ProjectDetailStyle::create($projectdetailstyle);
+			}
+			$projectdetail['projectdetailstyles'] = $projectdetailstyles;
+
+			return $projectdetail;
     }
 
     public function destroy($id)
     {
+      if(Account::checkOwner($id)==false && Account::checkTeamManager($id)==false) return;
+
         $projectdetailstyles = ProjectDetailStyle::where('projectdetailid', $id)->get();
         foreach($projectdetailstyles as $projectdetailstyle) $projectdetailstyle->delete();
         $projectdetail = ProjectDetail::find($id)->first();
